Move FakeSelectUserPresenter to dedicated module
